fix: Add SEA support directly to vendored node-pty and rename to node-pty
- Rename vendored-pty directory to node-pty for consistency
- Add built-in SEA (Single Executable Application) support to unixTerminal.ts
- Remove patching logic from build-native.js as SEA support is now native
- Update macOS deployment target to 14.0 in binding.gyp
- Fix native executable build issues by using process.dlopen in SEA mode
