+++ web/build-native.js
@@ -79,148 +79,12 @@ process.on('SIGTERM', () => {
   process.exit(1);
 });
 
-function applyMinimalPatches() {
-  console.log('Applying minimal SEA patches to node-pty...');
-  
-  // Create sea-loader.js
-  const seaLoaderPath = path.join(__dirname, 'node_modules/node-pty/lib/sea-loader.js');
-  if (!fs.existsSync(seaLoaderPath)) {
-    const seaLoaderContent = `"use strict";
-/* VIBETUNNEL_SEA_LOADER */
-Object.defineProperty(exports, "__esModule", { value: true });
-var path = require("path");
-var fs = require("fs");
-
-// Custom loader for SEA that uses process.dlopen
-var pty;
-
-// Helper function to load native module using dlopen
-function loadNativeModule(modulePath) {
-  const module = { exports: {} };
-  process.dlopen(module, modulePath);
-  return module.exports;
-}
-
-// Determine the path to pty.node
-function getPtyPath() {
-  const execDir = path.dirname(process.execPath);
-  // Look for pty.node next to the executable first
-  const ptyPath = path.join(execDir, 'pty.node');
-
-  if (fs.existsSync(ptyPath)) {
-    // Add path validation for security
-    const resolvedPath = path.resolve(ptyPath);
-    const resolvedExecDir = path.resolve(execDir);
-    if (!resolvedPath.startsWith(resolvedExecDir)) {
-      throw new Error('Invalid pty.node path detected');
-    }
-    return ptyPath;
-  }
-
-  // If not found, throw error with helpful message
-  throw new Error('Could not find pty.node next to executable at: ' + ptyPath);
-}
-
-try {
-  const ptyPath = getPtyPath();
-
-  // Set spawn-helper path for macOS only
-  // Linux uses forkpty() directly and doesn't need spawn-helper
-  if (process.platform === 'darwin') {
-    const execDir = path.dirname(process.execPath);
-    const spawnHelperPath = path.join(execDir, 'spawn-helper');
-    if (fs.existsSync(spawnHelperPath)) {
-      process.env.NODE_PTY_SPAWN_HELPER_PATH = spawnHelperPath;
-    }
-  }
-
-  pty = loadNativeModule(ptyPath);
-} catch (error) {
-  console.error('Failed to load pty.node:', error);
-  throw error;
-}
-
-exports.default = pty;
-`;
-    fs.writeFileSync(seaLoaderPath, seaLoaderContent);
-  }
-  
-  // Patch index.js
-  const indexPath = path.join(__dirname, 'node_modules/node-pty/lib/index.js');
-  if (fs.existsSync(indexPath)) {
-    let content = fs.readFileSync(indexPath, 'utf8');
-    if (!content.includes('VIBETUNNEL_SEA')) {
-      content = content.replace(
-        "exports.native = (process.platform !== 'win32' ? require('../build/Release/pty.node') : null);",
-        "exports.native = (process.platform !== 'win32' ? (process.env.VIBETUNNEL_SEA ? require('./sea-loader').default : require('../build/Release/pty.node')) : null);"
-      );
-      fs.writeFileSync(indexPath, content);
-    }
-  }
-  
-  // Patch unixTerminal.js
-  const unixPath = path.join(__dirname, 'node_modules/node-pty/lib/unixTerminal.js');
-  if (fs.existsSync(unixPath)) {
-    let content = fs.readFileSync(unixPath, 'utf8');
-    if (!content.includes('VIBETUNNEL_SEA')) {
-      // Find and replace the pty loading section
-      const startMarker = 'var pty;\nvar helperPath;';
-      const endMarker = 'var DEFAULT_FILE = \'sh\';';
-      const startIdx = content.indexOf(startMarker);
-      const endIdx = content.indexOf(endMarker);
-      
-      if (startIdx !== -1 && endIdx !== -1) {
-        const newSection = `var pty;
-var helperPath;
-// For SEA, check environment variables
-if (process.env.VIBETUNNEL_SEA) {
-    pty = require('./sea-loader').default;
-    // In SEA context, look for spawn-helper on macOS only (Linux doesn't use it)
-    if (process.platform === 'darwin') {
-        const execDir = path.dirname(process.execPath);
-        const spawnHelperPath = path.join(execDir, 'spawn-helper');
-        if (require('fs').existsSync(spawnHelperPath)) {
-            helperPath = spawnHelperPath;
-        } else if (process.env.NODE_PTY_SPAWN_HELPER_PATH) {
-            helperPath = process.env.NODE_PTY_SPAWN_HELPER_PATH;
-        }
-    }
-    // On Linux, helperPath remains undefined which is fine
-} else {
-    // Original loading logic
-    try {
-        pty = require('../build/Release/pty.node');
-        helperPath = '../build/Release/spawn-helper';
-    }
-    catch (outerError) {
-        try {
-            pty = require('../build/Debug/pty.node');
-            helperPath = '../build/Debug/spawn-helper';
-        }
-        catch (innerError) {
-            console.error('innerError', innerError);
-            // Re-throw the exception from the Release require if the Debug require fails as well
-            throw outerError;
-        }
-    }
-    helperPath = path.resolve(__dirname, helperPath);
-    helperPath = helperPath.replace('app.asar', 'app.asar.unpacked');
-    helperPath = helperPath.replace('node_modules.asar', 'node_modules.asar.unpacked');
-}
-`;
-        content = content.substring(0, startIdx) + newSection + content.substring(endIdx);
-        fs.writeFileSync(unixPath, content);
-      }
-    }
-  }
-  
-  console.log('SEA patches applied successfully');
-}
+// No patching needed - SEA support is built into our vendored node-pty
 
 async function main() {
   try {
-    // Apply minimal patches to node-pty
-    applyMinimalPatches();
+    // No patching needed - SEA support is built into our vendored node-pty
+    console.log('Using vendored node-pty with built-in SEA support...');
     
     // Ensure native modules are built (in case postinstall didn't run)
     const nativePtyDir = 'node_modules/node-pty/build/Release';
