+++ web/node-pty/binding.gyp
@@ -11,7 +11,7 @@
     'xcode_settings': {
       'GCC_ENABLE_CPP_EXCEPTIONS': 'YES',
       'CLANG_CXX_LIBRARY': 'libc++',
-      'MACOSX_DEPLOYMENT_TARGET': '10.7',
+      'MACOSX_DEPLOYMENT_TARGET': '14.0',
     },
     'msvs_settings': {
       'VCCLCompilerTool': { 'ExceptionHandling': 1 },
@@ -43,7 +43,7 @@
         'conditions': [
           ['OS=="mac"', {
             'xcode_settings': {
-              'MACOSX_DEPLOYMENT_TARGET': '10.12'
+              'MACOSX_DEPLOYMENT_TARGET': '14.0'
             }
           }]
         ]
