+++ web/node-pty/src/eventEmitter2.ts
