+++ web/node-pty/src/interfaces.ts
