+++ web/node-pty/src/native.d.ts
