+++ web/node-pty/src/utils.ts
