--- web/vendored-pty/lib/types.d.ts
@@ -1,12 +1,0 @@
-/**
- * Copyright (c) 2017, Daniel Imms (MIT License).
- * Copyright (c) 2018, Microsoft Corporation (MIT License).
- */
-export type ArgvOrCommandLine = string[] | string;
-export interface IExitEvent {
-    exitCode: number;
-    signal: number | undefined;
-}
-export interface IDisposable {
-    dispose(): void;
-}
