--- web/vendored-pty/lib/types.js
@@ -1,6 +1,0 @@
-"use strict";
-/**
- * Copyright (c) 2017, Daniel Imms (MIT License).
- * Copyright (c) 2018, Microsoft Corporation (MIT License).
- */
-Object.defineProperty(exports, "__esModule", { value: true });
