--- web/vendored-pty/lib/utils.d.ts
@@ -1,5 +1,0 @@
-/**
- * Copyright (c) 2017, Daniel Imms (MIT License).
- * Copyright (c) 2018, Microsoft Corporation (MIT License).
- */
-export declare function assign(target: any, ...sources: any[]): any;
